refac: padroniza tipo de dado do ID para long
* fix(cliente): padroniza tipo de dado de ID par a long

* fix(fornecedor): padroniza tipo de dado de ID par a long

* refactor(veiculo): move entidade veiculo para package dedicado

* fix(venda): padroniza tipo de dado de ID par a long

+++ src/model/entity/Venda.java
@@ -9,8 +9,8 @@ import java.time.LocalDateTime;
 
 public class Venda {
 
-    private final int id;
-    private static int idContador = 1;
+    private final long id;
+    private static long idContador = 1;
     private final LocalDateTime dataHora;
     private Bomba bomba;
     private Funcionario funcionario;
@@ -31,7 +31,7 @@ public class Venda {
     // |----------| Getters e Setters |----------|
 
 
-    public int getId() {
+    public long getId() {
         return id;
     }
